rebuild5/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
